Extract from statements from sparql and use that to initialize model.
the librdf python bindings don't support from <url> so, manually
extract it. This makes the query format look nicer than having one
line with a really long list of urls on it.

Also I changed my test code into unittests instead of just
a running to see if it works.